% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{class_plot}
\alias{class_plot}
\title{Classification plot}
\usage{
class_plot(classification_mat = NULL, directory = NULL, n = NULL)
}
\arguments{
\item{classification_mat}{data.frame; see 'classification' function}

\item{directory}{string; output directory}

\item{n}{numeric; number of top n genes ranked by mND to plot}
}
\description{
Visulation of classification
}
