% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Gene classification}
\usage{
classification(mND, X0, Hl, top = NULL, alpha = NULL)
}
\arguments{
\item{mND}{mND scores; if data were permuted use the output of 'signif_assess' function, otherwise output of 'mND' function}

\item{X0}{matrix; initial score matrix X0, in which each column (layer) represents a score vector over all vertices of G}

\item{Hl}{list; defined by the user, composed of high scoring genes names in each layer of X0.}

\item{top}{array; number of genes with the highest neighbourhoods to define the gene set Nl (see below); 1 number for each layer.}

\item{alpha}{array; significance level on the empirical p-value to define the gene set Nl (see below); 1 number for each layer.}
}
\value{
list with:
\itemize{
\item{\code{gene_class}}{: data.frame with classifications of each gene in every layers.
The possible classification are: M (Module); L (Linker); I (Isolated); NS (Not Significant);}
\item{\code{occ_labels}}{: occurrence of the different labels (M; L; I; NS) for each gene across layers.}
}
}
\description{
Gene classification: genes are classified in every layer
}
\details{
The gene set Nl contains the genes with the highest neighbourhoods in layer l. If both 'top' and 'alpha' are provided, a maximum of top genes with p < alpha will be considered.
}
