% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleaning_mnd}
\alias{cleaning_mnd}
\title{Create a dataset for survival and incidence calculation}
\usage{
cleaning_mnd(demo, dx, rx, codes_sys, riluzole_name = "riluzole|riluteck", ...)
}
\arguments{
\item{demo}{the dataset with demographic information, including id, dob, dod, sex, onset_date. Pls check the data shall.}

\item{dx}{the dataset with all diagnosis information, including id, codes, ref_date, setting. Pls check the data shall.}

\item{rx}{the dataset with all prescription records, including id, drug name, date of prescription start and end, type of presciption (IP, OP, AE, Discharge). Pls check the data shall.}

\item{codes_sys}{can be "icd9", "icd10", or "readcodes"}

\item{riluzole_name}{the name in your database stands for riluzole.}
}
\value{
two dataset:1) for incidence 2) for time varing cox
}
\description{
Create a dataset for survival and incidence calculation
}
\examples{
cleaning_mnd(demo, dx, codes_sys="icd9")
}
