% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_k.R
\name{eval_k}
\alias{eval_k}
\title{Evaluation of the parameter k}
\usage{
eval_k(mND, X0, G, top = 200)
}
\arguments{
\item{mND}{list; list of mND score obtained for different k values, list components should be named as the k values used}

\item{X0}{matrix; initial score matrix X0, in which each column (layer) represents a score vector over all vertices of G}

\item{G}{graph; gene x gene undirected interaction graph}

\item{top}{numeric; number of genes to be considered at the top of the list of genes ranked by mND}
}
\value{
list with calculation of w at varying k and plot with results.
}
\description{
Evaluation of the parameter k
}
