% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework_mND.R
\name{framework_mND}
\alias{framework_mND}
\title{mND global function}
\usage{
framework_mND(
  X0 = NULL,
  W = NULL,
  alpha = 0.7,
  nMax = 10000,
  eps = 1e-06,
  k = 3,
  r = 0,
  seed_n = NULL,
  BPPARAM = NULL,
  ...
)
}
\arguments{
\item{X0}{matrix; each column (layer) of the matrix X0 is a score vector over all vertices of G.}

\item{W}{matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}

\item{alpha}{numeric; the smothing factor}

\item{nMax}{numeric; maximum number of iterations}

\item{eps}{numeric; the iteration will stop when the maximum difference between matrix Xs between two consecutive iteraction is smaller than \code{eps}}

\item{k}{numeric; number of top k first neighbors}

\item{r}{numeric; number of permutations}

\item{seed_n}{numeric; a given 'seed_n' will determine the same permutations of X0 and ensures the reproducibility of the analysis}

\item{BPPARAM}{optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(). See ?bplapply.}

\item{...}{additional parameter to NPATools::perm_vertices()}
}
\value{
list with:
\itemize{
\item{\code{mND}}{: mND score (mND); if r parameter was used, the following columns will be added: corrisponding empirical p-value (p), product of mND score and -log10(p) (mNDp)}
\item{\code{t}}{: sum of top k first neighbours}
\item{\code{pl}}{: corrisponding empirical p-values of (t), only if r parameter was used}
\item{\code{tp}}{: product of the sum of top k neighbours (t) and -log10(pl), only if r parameter was used}
}
}
\description{
It calculates permutations of X0 (if r parameter is used), applies network-diffusion on data, computes the mND score and the relative empirical p-value (if data are permuted).
Outputs of function can be used to classify genes in each layer with the classification function.
}
