% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_DT_Exposure_Endpoint}
\alias{get_DT_Exposure_Endpoint}
\title{Combine the datasets with exposure and outcomes}
\usage{
get_DT_Exposure_Endpoint(
  demo,
  rx,
  ip,
  riluzole_name = "riluzole|riluteck",
  ...
)
}
\arguments{
\item{demo}{the dataset with demographic information, including id, dob, dod, sex, onset_date. Pls check the data shall.}

\item{rx}{the dataset with all prescription records, including id, drug name, date of prescription start and end, type of presciption (IP, OP, AE, Discharge). Pls check the data shall.}

\item{ip}{the dataset with all in-hospitalization records, including id, date of adminssion, date of discharge, type of records (setting, IP, OP, AE). Pls check the data shall.}

\item{riluzole_name}{the name in your database stands for riluzole.}
}
\value{
a combined data set in data table format. One patients may have multiple rows of records.
}
\description{
Combine the exposure (Riluzole history) and outcome (Hospitalization records) for SCCS study design.
}
\examples{
get_DT_Exposure_Endpoint(demo,ip,rx)
            id sex        dob        dod onset_date date_rx_st date_rx_end   date_adm   date_dsg
   1: 10038681   M 1966-09-10 2017-06-06 2015-07-23 2013-05-31  2014-06-02 2015-07-21 2015-07-23
   2: 10038681   M 1966-09-10 2017-06-06 2015-07-23 2013-05-31  2014-06-02 2016-10-17 2016-10-22
   3: 10038681   M 1966-09-10 2017-06-06 2015-07-23 2013-05-31  2014-06-02 2017-01-24 2017-01-27
   4: 10038681   M 1966-09-10 2017-06-06 2015-07-23 2013-05-31  2014-06-02 2017-03-03 2017-03-31
   5: 10038681   M 1966-09-10 2017-06-06 2015-07-23 2013-05-31  2014-06-02 2017-04-10 2017-04-10
---
2663:  9903734   F 1955-07-02 2018-11-28 2017-11-28 2018-05-28  2018-07-22 2018-07-30 2018-08-10
2664:  9903734   F 1955-07-02 2018-11-28 2017-11-28 2018-05-28  2018-07-22 2018-10-31 2018-11-02
2665:  9903734   F 1955-07-02 2018-11-28 2017-11-28 2018-07-25  2018-11-20 2017-11-27 2017-11-28
2666:  9903734   F 1955-07-02 2018-11-28 2017-11-28 2018-07-25  2018-11-20 2018-07-30 2018-08-10
2667:  9903734   F 1955-07-02 2018-11-28 2017-11-28 2018-07-25  2018-11-20 2018-10-31 2018-11-02
}
