% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_DT_SCCS}
\alias{get_DT_SCCS}
\title{Genrate the dataset and Cleaning for SCCS}
\usage{
get_DT_SCCS(data, obst, obed, ...)
}
\arguments{
\item{data}{the dataset including exposure and endpoint information}

\item{obst}{the defined study observation date. Choose the earliest avaiable date with *Riluzole* in the hospital or approved date by FDA. For instance, 2001-8-24 in Hong Kong.}
}
\value{
a dataset with cutted time periods for sccs study design
}
\description{
Genrate the dataset and Cleaning for SCCS
}
\examples{
no example
}
