% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_DT_SCCS_collapsed}
\alias{get_DT_SCCS_collapsed}
\title{Data created for collapsed analysis}
\usage{
get_DT_SCCS_collapsed(data, obst, obed, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Data created for collapsed analysis
}
