% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_px_dx}
\alias{get_px_dx}
\title{Obtain the past hx}
\usage{
get_px_dx(data, dx, icd9)
}
\arguments{
\item{data}{the data after initial cleanning}

\item{dx}{the dataset with all diagnosis information, including id, codes, ref_date, setting. Pls check the data shall.}

\item{icd9}{}
}
\value{

}
\description{
Obtain the past hx
}
\examples{
get_px_dx(database,dx, icd9)
}
