% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{get_tableone}
\alias{get_tableone}
\title{Generate table one for the cohort}
\usage{
get_tableone(x)
}
\arguments{
\item{x}{dataset including data after cleaning}
}
\value{

}
\description{
Generate table one for the cohort
}
\examples{
get_tableone(x)
}
