% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_adj.R
\docType{data}
\name{ind_adj}
\alias{ind_adj}
\title{ind_adj,  index of neighbour for each genes using STRING interactome}
\format{
list
}
\usage{
ind_adj
}
\description{
ind_adj,  index of neighbour for each genes using STRING interactome
}
\keyword{datasets}
