% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mND.R
\name{mND}
\alias{mND}
\title{mND score}
\usage{
mND(Xs = NULL, ind_adj = NULL, k = 3, BPPARAM = NULL, min.norm.val = 0)
}
\arguments{
\item{Xs}{a matrix or list of matrices (if X0 was permuted, where the first element of the list is the one obtained with real data) with network diffusion scores, see 'ND' function}

\item{ind_adj}{list; index of neighbours obtained from adjacency matrix, see 'neighbour_index' function}

\item{k}{numeric; number of top k first neighbors}

\item{BPPARAM}{optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(). See ?bplapply.}

\item{min.norm.val}{normalized ND values below this value will be set to 0; this is useful to denoise the steady state of ND and focus only on the highest scores}
}
\value{
list with:
\itemize{
\item{\code{mND}}{: mND score}
\item{\code{t}}{: sum of top k neighbours}
}
If data were permuted (therefore Xs is a list), the output list described above will be reported for each permutation of Xs.
}
\description{
Calculation of mND score
}
