% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mND_s.R
\name{mND_s}
\alias{mND_s}
\title{Integration with mND function}
\usage{
mND_s(output = NULL, ind_adj = NULL, k = NULL, min.norm.val = 0)
}
\arguments{
\item{output}{Xs}

\item{ind_adj}{index of neighbours obtained from adj matrix, see 'neighbour_index' function}

\item{k}{number of top k first neighbors}

\item{min.norm.val}{normalized ND values below this value will be set to 0; this is useful to denoise the steady state of ND and focus only on the highest scores}
}
\value{
a list with:
\itemize{
\item{\code{mND}}{: mND score}
\item{\code{t}}{: sum of top k neighbours}
}
}
\description{
Integration with mND function
}
