% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND.R
\name{ND}
\alias{ND}
\title{Network diffusion}
\usage{
ND(X0, W, alpha = 0.7, nMax = 10000, eps = 1e-06, cores = 1)
}
\arguments{
\item{X0}{matrix or a list of matrices (if X0 was permuted (see 'perm_X0' function), where the first element of the list is the one obtained with real data); each column (layer) of the matrix X0 is a score vector over all vertices of G.}

\item{W}{matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}

\item{alpha}{numeric; the smothing factor}

\item{nMax}{numeric; maximum number of iterations}

\item{eps}{numeric; the iteration will stop when the maximum difference between matrix Xs between two consecutive iteraction is smaller than \code{eps}}

\item{cores}{numeric; number of cores to run in parallel in case of permutations}
}
\value{
a matrix or a list of matrices (if data were permuted) with network diffusion scores.
}
\description{
Network diffusion: input scores (X0) are smoothed by network-diffusion, obtaining the corresponding network-constrained scores Xs.
}
