% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND_s.R
\name{ND_s}
\alias{ND_s}
\title{Network Diffusion using}
\usage{
ND_s(
  X0,
  W,
  alpha = 0.7,
  nMax = 10000,
  eps = 1e-06,
  final.smooth = FALSE,
  all.steps = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X0}{matrix composed of column vectors with initial distribution of information}

\item{W}{symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}

\item{alpha}{numeric, the smothing factor}

\item{nMax}{numeric, maximum number of iterations}

\item{eps}{numeric, the iteration will stop when the maximum difference between matrix Xs between two consecutive iteraction is smaller than \code{eps}}

\item{final.smooth}{TRUE/FALSE, whether to do the final step of smoothing}

\item{all.steps, }{TRUE/FALSE, whether to store all steps}

\item{verbose, }{TRUE/FALSE}
}
\value{
a list with:
\itemize{
\item{\code{Xt}}{ the smoothed matrix;}
\item{\code{eps}}{ see above;}
\item{\code{max.abs.diff}}{ max(abs(F_t) - abs(F_{t-1}));}
\item{\code{Xs.all}}{ transient Xs matrices.}
}
}
\description{
Network Diffusion using
}
