% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_index.R
\name{neighbour_index}
\alias{neighbour_index}
\title{Index neighbours}
\usage{
neighbour_index(W = NULL)
}
\arguments{
\item{W}{matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}
}
\value{
list with indices of neighbours for each gene
}
\description{
Indices of neighbours for each gene
}
