% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omega}
\alias{omega}
\title{Omega}
\usage{
omega(G, u)
}
\arguments{
\item{G}{graph; gene x gene undirected interaction graph}

\item{u}{list; ranked names that have to be those in V(G)$name}
}
\description{
Calculation of omega function
}
