% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{p_inci}
\alias{p_inci}
\title{Plot the figure of incidence}
\usage{
p_inci(data, region = "Hong Kong")
}
\arguments{
\item{data}{the dataset generated from run_incidence}

\item{region}{the site name which will be shown in the plot}
}
\value{

}
\description{
Plot the figure of incidence
}
\examples{
p_inci(data$std_inci)
}
