% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{p_inci_sex}
\alias{p_inci_sex}
\title{Plot the figure of incidence by age}
\usage{
p_inci_sex(data, region = "Hong Kong")
}
\arguments{
\item{data}{the dataset generated from run_incidence}
}
\value{

}
\description{
Plot the figure of incidence by age
}
\examples{
p_inci_sex(data)
}
