% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{p_inci_type}
\alias{p_inci_type}
\title{Plot the figure of incidence by subgroup}
\usage{
p_inci_type(data, region = "Hong Kong")
}
\arguments{
\item{data}{the dateset generated from run_incidence}

\item{region}{the site name which will be shown in the plot}
}
\value{

}
\description{
Plot the figure of incidence by subgroup
}
\examples{
p_inci_type(data)
}
