% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_X0.R
\name{perm_X0}
\alias{perm_X0}
\title{Permutation of X0}
\usage{
perm_X0(X0, r, W, seed_n = NULL)
}
\arguments{
\item{X0}{matrix; a score matrix X0, in which each column (layer) is a score vector over all vertices of G.}

\item{r}{numeric; number of permutations.}

\item{W}{matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}

\item{seed_n}{optional numeric; to specify the seed for (pseudo) random number generation; using the same seed}
}
\value{
list with permutations of X0 (where the first element of the list is the one obtained with real data)
}
\description{
Permutation of input matrix X0
}
