% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(
  mND = NULL,
  classification = NULL,
  W = NULL,
  n = NULL,
  directory = NULL
)
}
\arguments{
\item{mND}{mND scores; if data were permuted use the output of 'signif_assess' function, otherwise output of 'mND' function.}

\item{classification}{data.frame; gene classification, output of 'classification' function}

\item{W}{matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}

\item{n}{numeric; number of top n ranking genes to consider in the plot of gene networks}

\item{directory}{string; output directory}
}
\description{
Visualization of results.
}
\details{
The possible plots are:
\itemize{
\item{\code{genes ranked by mND score and the corresponding p-value (for permutated data)}}{: mND argument is required (output of 'signif_assess' function);}
\item{\code{gene networks composed of the top n ranking genes}}{: mND, W and n arguments are required;}
\item{\code{gene classification for the top 100 ranking genes across layers}}{: classification argument is required. Legend: brown: isolated; orange: linker; purple: module; grey: not significant.}
}
}
