% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{print.mndinci}
\alias{print.mndinci}
\title{print only std inci, not raw data}
\usage{
\method{print}{mndinci}(x)
}
\arguments{
\item{x}{the incidence result}
}
\value{

}
\description{
print only std inci, not raw data
}
