% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{print.mndsccsoutput}
\alias{print.mndsccsoutput}
\title{show only the primary result}
\usage{
\method{print}{mndsccsoutput}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
show only the primary result
}
