% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{saveall}
\alias{saveall}
\title{Title}
\usage{
saveall(output_desc, output_sccs, dir_output = ".")
}
\arguments{
\item{output_desc}{the output from run_desc()}

\item{output_sccs}{the output from run_sccs()}

\item{dir_ouput}{the file directory that you want to use}
}
\value{

}
\description{
Title
}
