% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{sccs}
\alias{sccs}
\title{Standardsccs with formated output}
\usage{
sccs(fml, dob13, ...)
}
\arguments{
\item{fml}{the formula in standardsccs}
}
\value{

}
\description{
Standardsccs with formated output
}
\examples{
sccs()
}
