% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_assess.R
\name{signif_assess}
\alias{signif_assess}
\title{Significance assessment}
\usage{
signif_assess(mND)
}
\arguments{
\item{mND}{list; output of 'mND' function}
}
\value{
list with:
\itemize{
\item{\code{mND}}{: mND score (mND), corrisponding empirical p-value (p), product of mND score and -log10(p) (mNDp)}
\item{\code{t}}{: sum of top k first neighbours}
\item{\code{pl}}{: corrisponding empirical p-values of (t)}
\item{\code{tp}}{: product of the sum of top k neighbours (t) and -log10(pl)}
}
}
\description{
Significance assessment: empirical p-values are calculated by using the pool of r permuted versions of X0 (clearly, use this function only if 'perm_X0' function was used).
}
