% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{sxd}
\alias{sxd}
\title{Print with rounding, with 0 at the end if appropriate}
\usage{
sxd(x, n = 2)
}
\arguments{
\item{x}{number}

\item{n}{digits in round}
}
\value{

}
\description{
Print with rounding, with 0 at the end if appropriate
}
\examples{
sxd(2.01999,2)
}
