% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_4_survival}
\alias{clean_4_survival}
\title{Create a dataset for survival and incidence calculation}
\usage{
clean_4_survival(demo, dx, codes_sys, ...)
}
\arguments{
\item{codes_sys}{can be "icd9", "icd10", or "readcodes"}
}
\value{

}
\description{
Create a dataset for survival and incidence calculation
}
