% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mND.R
\name{mND}
\alias{mND}
\title{mND score}
\usage{
mND(Xs, ind_adj, k = 3, cores = 1)
}
\arguments{
\item{Xs}{a matrix or list of matrices (if X0 was permuted, where the first element of the list is the one obtained with real data) with network diffusion scores, see 'ND' function}

\item{ind_adj}{list; index of neighbours obtained from adjacency matrix, see 'neighbour_index' function}

\item{k}{numeric; number of top k first neighbors}

\item{cores}{numeric; number of cores to run in parallel in case of permutations}
}
\value{
list with:
\itemize{
\item{\code{mND}}{: mND score}
\item{\code{t}}{: sum of top k neighbours}
}
If data were permuted (therefore Xs is a list), the output list described above will be reported for each permutation of Xs.
}
\description{
Calculation of mND score
}
