% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mND_s.R
\name{mND_s}
\alias{mND_s}
\title{Integration with mND function}
\usage{
mND_s(output, ind_adj, k)
}
\arguments{
\item{output}{Xs}

\item{ind_adj}{index of neighbours obtained from adj matrix, see 'neighbour_index' function}

\item{k}{number of top k first neighbors}
}
\value{
a list with:
\itemize{
\item{\code{mND}}{: mND score}
\item{\code{t}}{: sum of top k neighbours}
}
}
\description{
Integration with mND function
}
