% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_k.R
\name{optimize_k}
\alias{optimize_k}
\title{Optimization of k value}
\usage{
optimize_k(Xs, X0, k, ind_adj, W, top = 200, cores = 4)
}
\arguments{
\item{Xs}{a matrix or list of matrices (if X0 was permuted, where the first element of the list is the one obtained with real data) with network diffusion scores, see 'ND' function}

\item{X0}{matrix; initial score matrix X0, in which each column (layer) represents a score vector over all vertices of G}

\item{k}{array; array with k values}

\item{ind_adj}{list; index of neighbours obtained from adjacency matrix, see 'neighbour_index' function}

\item{W}{matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function}

\item{top}{numeric; number of genes to be considered at the top of the list of genes ranked by mND}

\item{cores}{numeric; number of cores to run in parallel in case of permutations}
}
\value{
list with calculation of w at varying k and plot with results.
}
\description{
Optimization of the 'k' value by selecting a value that yields connected networks enriched in initial scores.
}
