% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_incidence}
\alias{run_incidence}
\title{run analysis for incidence}
\usage{
run_incidence(demo, dx, region = "hk", codes_sys = "icd9")
}
\arguments{
\item{codes_sys}{}
}
\value{

}
\description{
run analysis for incidence
}
