% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx_clean.R
\name{shrink_interval}
\alias{shrink_interval}
\title{Periods cleaning}
\usage{
shrink_interval(data, st, ed, gap = 1)
}
\arguments{
\item{data}{the prescription dataset}

\item{st}{the column name for starting date}

\item{ed}{the column name for ending date}

\item{gap}{the gap between two prescription records which you would like to merge. For instance, the two Rx records whith interval less than 7 days may be consider as the same prescription.}
}
\value{

}
\description{
merge the overlapping or closed records, eg.prescription records, hospitalization
}
\examples{
data <- data.table(id=rep(1,2),
        date_st=lubridate::ymd(c("2021-1-1","2021-1-4")),
        date_end=lubridate::ymd(c("2021-1-5","2021-1-6")))
shrink_interval(data,"date_st","date_end",gap=2)
}
